groov-server/app/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
Django==5.1.3
dkimpy==1.1.8
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.5
filelock==3.16.1
Flask==3.0.3
fonttools==4.55.3
frozenlist==1.4.1
fsspec==2024.10.0
geoip2==4.8.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
h11==0.14.0
httplib2==0.22.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jwt==1.3.1
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.10.0
maxminddb==2.6.2
mpmath==1.3.0
multidict==6.1.0
mutagen==1.47.0
mysql==0.0.3
mysqlclient==2.2.6
networkx==3.4.2
norminette==3.3.51
numpy==2.1.1
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
PyMySQL==1.1.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
python-spf==0.0.1
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
sqlparse==0.5.2
starlette==0.41.3
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
waitress==3.0.0
Werkzeug==3.0.4
yarl==1.11.1
